+++ substrafl/algorithms/pytorch/torch_newton_raphson_algo.py
@@ -267,7 +267,7 @@ class TorchNewtonRaphsonAlgo(TorchAlgo):
         self._model.eval()
 
         predictions = torch.Tensor([]).to(self._device)
-        with torch.inference_mode():
+        with torch.no_grad():
             for x in predict_loader:
                 x = x.to(self._device)
                 predictions = torch.cat((predictions, self._model(x)), 0)