+++ substrafl/strategies/fed_avg.py
@@ -101,8 +101,8 @@ class FedAvg(Strategy):
             clean_models (bool): Clean the intermediary models of this round on the Substra platform.
                 Set it to False if you want to download or re-use intermediary models. This causes the disk
                 space to fill quickly so should be set to True unless needed.
-            additional_orgs_permissions (typing.Optional[set]): Additional permissions to give to the model outputs
-                after training, in order to test the model on an other organization.
+            additional_orgs_permissions (typing.Optional[set]): Additional permissions to give to the model
+                outputs after training, in order to test the model on an other organization.
         """
         if aggregation_node is None:
             raise ValueError("In FedAvg strategy aggregation node cannot be None")